% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore a dataframe}
\usage{
explore(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A list with output from class, colnames, dim and summary
}
\description{
Explore a dataframe
}
\examples{
data <- data.frame(points=c(3, 2, 4, 3, 4, 8),
assists=c(6, NA, 7, NA, 7, 9),
team=c('A', 'A', 'A', NA, 'C', 'C'))
explore(data)
}
