% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAcol.R
\name{NAcol}
\alias{NAcol}
\title{Find how many NA in a column}
\usage{
NAcol(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
How many NA's in a column
}
\description{
Find how many NA in a column
}
\examples{
data <- data.frame(points=c(3, 2, 4, 3, 4, 8),
assists=c(6, NA, 7, NA, 7, 9),
team=c('A', 'A', 'A', NA, 'C', 'C'))
explore(data)
}
