% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NArow.R
\name{NArow}
\alias{NArow}
\title{Find how many NA in a row if there are any}
\usage{
NArow(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
How many NA's in a row
}
\description{
Find how many NA in a row if there are any
}
\examples{
data <- data.frame(points=c(3, 2, 4, 3, 4, 8),
assists=c(6, NA, 7, NA, 7, 9),
team=c('A', 'A', 'A', NA, 'C', 'C'))
explore(data)
}
