% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace}
\alias{replace}
\title{Replaces value to new value in specific column of a dataframe}
\usage{
replace(data, column, oldString, newString)
}
\arguments{
\item{data}{A dataframe}

\item{column}{A columnname as.character of the dataframe}

\item{oldString}{Old value}

\item{newString}{New value}
}
\value{
Replaced dataframe
}
\description{
Replaces value to new value in specific column of a dataframe
}
