% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soccer}
\alias{soccer}
\title{Soccer dataset.}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{points}{points scored}
\item{assists}{assists made}
\item{team}{teams played}
}
}
\source{
\url{https://www.statology.org/replace-values-in-data-frame-r/}
}
\usage{
soccer
}
\description{
A dataset containing the points scored, assists made and teams played of a made-up soccer player.
}
\keyword{datasets}
